Reorganize singleLineString into common utilities
